SNC-6470. Build issues on Qt6

Closes SNC-6470

See merge request desktop/desktop!4289

+++ src/MEGASync/node_selector/model/NodeSelectorModel.cpp
@@ -93,7 +93,7 @@ void NodeRequester::requestNodeAndCreateChildren(NodeSelectorModelItem* item,
                 lockDataMutex(true);
                 item->initializeChildItems(childItems);
                 lockDataMutex(false);
-                emit nodesReady(item, childCount);
+                emit nodesReady(item, static_cast<int>(childCount));
             }
         }
     }
